refactor: move `can` to `utils.permission`
